requirements.txt
Flask==3.0.0
flask-login==0.6.3
python-dotenv==1.0.0
requests==2.31.0
oauthlib==3.2.2
pyOpenSSL==23.3.0
squareup==33.0.0.20231115
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
email-validator==2.1.0.post1
Werkzeug==3.0.1
alembic==1.13.1
SQLAlchemy==2.0.25
python-json-logger==2.0.7
